% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityLearner.R
\name{LearnerDensity}
\alias{LearnerDensity}
\title{Density Estimation Learner}
\description{
This learner implements the \code{\link[mlr3:Learner]{mlr3::Learner}} for density estimation \code{\link[mlr3:Task]{Task}}s.
\itemize{
\item \code{task_type} is set to \code{"density"}.
\item \verb{Creates [}Prediction\verb{][mlr3::Prediction]s of class [}PredictionDensity`].
\item \code{predict_types} always has the value \code{"prob"}.
}
}
\seealso{
Other density estimation classes: 
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super class}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{LearnerDensity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerDensity-new}{\code{LearnerDensity$new()}}
\item \href{#method-LearnerDensity-sample}{\code{LearnerDensity$sample()}}
\item \href{#method-LearnerDensity-clone}{\code{LearnerDensity$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensity-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{LearnerDensity} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$new(
  id,
  param_set = ps(),
  predict_types = "prob",
  feature_types = character(0),
  properties = character(0),
  data_formats = "data.table",
  packages = character(0),
  man = NA_character_
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensity-sample"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensity-sample}{}}}
\subsection{Method \code{sample()}}{
Sample from the fitted density.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$sample(n, lower = NULL, upper = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer(1)})\cr
Number of samples to generate.}

\item{\code{lower}}{(\code{numeric} | \code{NULL})\cr
lower bounds, to sample truncated distributions. Must contain \code{NA} for non-numeric features. \code{NULL} (default): do not truncate.}

\item{\code{upper}}{(\code{numeric} | \code{NULL})\cr
upper bounds, to sample truncated distributions. Must contain \code{NA} for non-numeric features. \code{NULL} (default): do not truncate.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
