% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityMeasure.R
\name{MeasureDensity}
\alias{MeasureDensity}
\title{Density Estimation Measure}
\description{
This measure implements \code{\link[mlr3:Measure]{mlr3::Measure}} for density estimation.
\itemize{
\item \code{task_type} is set to \code{"density"}.
\item \code{predict_type} must always be \code{"prob"}.
}
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super class}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{MeasureDensity}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureDensity-new}{\code{MeasureDensity$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureDensity-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureDensity-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{MeasureDensity} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureDensity$new(
  id,
  range,
  minimize = NA,
  aggregator = NULL,
  properties = character(0),
  predict_type = "prob",
  task_properties = character(0),
  packages = character(0),
  man = NA_character_
)}\if{html}{\out{</div>}}
}

}
}
