% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityLearner.R
\name{mlr_learners_density.featureless}
\alias{mlr_learners_density.featureless}
\alias{LearnerDensityFeatureless}
\title{Featureless Density Estimation Learner}
\description{
Simple \code{\link{LearnerDensity}} fallback learner which gives all predictions the density \code{.Machine$double.xmin} (typically \code{2.225074e-308}).

\verb{# TODO: could make this a hyperparameter.}
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_measures_density.logloss}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super classes}{
\code{\link[mlr3:Learner]{mlr3::Learner}} -> \code{\link[smashy:LearnerDensity]{smashy::LearnerDensity}} -> \code{LearnerDensityFeatureless}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerDensityFeatureless-new}{\code{LearnerDensityFeatureless$new()}}
\item \href{#method-LearnerDensityFeatureless-clone}{\code{LearnerDensityFeatureless$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="base_learner"><a href='../../mlr3/html/Learner.html#method-Learner-base_learner'><code>mlr3::Learner$base_learner()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="format"><a href='../../mlr3/html/Learner.html#method-Learner-format'><code>mlr3::Learner$format()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="help"><a href='../../mlr3/html/Learner.html#method-Learner-help'><code>mlr3::Learner$help()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict"><a href='../../mlr3/html/Learner.html#method-Learner-predict'><code>mlr3::Learner$predict()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="predict_newdata"><a href='../../mlr3/html/Learner.html#method-Learner-predict_newdata'><code>mlr3::Learner$predict_newdata()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="print"><a href='../../mlr3/html/Learner.html#method-Learner-print'><code>mlr3::Learner$print()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="reset"><a href='../../mlr3/html/Learner.html#method-Learner-reset'><code>mlr3::Learner$reset()</code></a></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Learner" data-id="train"><a href='../../mlr3/html/Learner.html#method-Learner-train'><code>mlr3::Learner$train()</code></a></li>
<li><span class="pkg-link" data-pkg="smashy" data-topic="LearnerDensity" data-id="sample"><a href='../../smashy/html/LearnerDensity.html#method-LearnerDensity-sample'><code>smashy::LearnerDensity$sample()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensityFeatureless-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensityFeatureless-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{LearnerDensityFeatureless} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensityFeatureless$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerDensityFeatureless-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerDensityFeatureless-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerDensityFeatureless$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
