% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityMeasure.R
\name{mlr_measures_density.logloss}
\alias{mlr_measures_density.logloss}
\alias{MeasureDensityLogloss}
\title{Density Logloss}
\description{
Logloss measure: Negative log likelihood of all predictions.
}
\seealso{
Other density estimation classes: 
\code{\link{LearnerDensity}},
\code{\link{MeasureDensity}},
\code{\link{PredictionDensity}},
\code{\link{TaskDensity}},
\code{\link{as_prediction_density}()},
\code{\link{as_task_density}()},
\code{\link{mlr_learners_density.featureless}},
\code{\link{mlr_learners_density.np}},
\code{\link{mlr_tasks_faithful}}
}
\concept{density estimation classes}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{smashy::MeasureDensity} -> \code{MeasureDensityLogloss}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureDensityLogloss-new}{\code{MeasureDensityLogloss$new()}}
\item \href{#method-MeasureDensityLogloss-clone}{\code{MeasureDensityLogloss$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureDensityLogloss-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureDensityLogloss-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureDensityLogloss$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureDensityLogloss-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureDensityLogloss-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureDensityLogloss$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
