% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerKD.R
\name{SamplerKD}
\alias{SamplerKD}
\title{KDE-model based Sampler}
\description{
Sample similar to BOHB

\verb{# TODO}

This currently hardcodes lots of things.
}
\section{Super class}{
\code{\link[paradox:Sampler]{paradox::Sampler}} -> \code{SamplerKD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SamplerKD-new}{\code{SamplerKD$new()}}
\item \href{#method-SamplerKD-clone}{\code{SamplerKD$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="paradox" data-topic="Sampler" data-id="format"><a href='../../paradox/html/Sampler.html#method-Sampler-format'><code>paradox::Sampler$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="paradox" data-topic="Sampler" data-id="print"><a href='../../paradox/html/Sampler.html#method-Sampler-print'><code>paradox::Sampler$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="paradox" data-topic="Sampler" data-id="sample"><a href='../../paradox/html/Sampler.html#method-Sampler-sample'><code>paradox::Sampler$sample()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplerKD-new"></a>}}
\if{latex}{\out{\hypertarget{method-SamplerKD-new}{}}}
\subsection{Method \code{new()}}{
Initialize the \code{SamplerKD} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamplerKD$new(
  param_set,
  task,
  minimize,
  alpha = 0.15,
  min_points_in_model = 0,
  bandwidth_factor = 3,
  min_bandwidth = 0.001
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{param_set}}{(\code{\link[paradox:ParamSet]{ParamSet}})\cr
\code{\link[paradox:ParamSet]{ParamSet}} to sample from.}

\item{\code{task}}{([\code{TaskRegr}][mlr3::TaskRegr)\cr
Regression \code{\link[mlr3:Task]{Task}} representing past performance values.}

\item{\code{minimize}}{(\code{logical})\cr
Whether outcome in \code{task} is to be minimized.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SamplerKD-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SamplerKD-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SamplerKD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
