% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{smashy-package}
\alias{smashy}
\alias{smashy-package}
\title{smashy: Synchronized Model Assisted Hyperband}
\description{
\code{smashy} offers both an \code{\link[bbotk:Optimizer]{Optimizer}} and a \code{\link[mlr3tuning:Tuner]{Tuner}} for general
smashy-optimization, but generally the \code{\link[=configure_smashy]{configure_smashy()}}-function should be used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mlr-org/smashy}
  \item Report bugs at \url{https://github.com/mlr-org/smashy/issues}
}

}
\author{
\strong{Maintainer}: Martin Binder \email{mlr.developer@mb706.com}

}
